+++ config/webpack.config.js
@@ -91,10 +91,7 @@ module.exports = ( env, argv ) => {
 						{
 							loader: `sass-loader`,
 							options: {
-								api: `modern`,
-								sassOptions: {
-									silenceDeprecations: [`mixed-decls`]
-								}
+								api: `modern`
 							}
 						}
 					]