+++ js/scripts.js
@@ -1,5 +1,3 @@
-// import im from 'include-media-export/include-media.js';
-
 ( function() {
 	// Remove the 'no-js' <body> class
 	document.body.classList.remove( `no-js` );